Update Rscript.R
Slight alteration to sine curves to make it work with Simon's redefinition of O

Changed the first section of the code to make it functional - preprocessing, finding w, finding u & v, and baseline correction are now all separate functions called in the main code.

There have been some major changed and fewer variables in the global environment - I think this code will NOT work if you try and run from start to end. Need to integrate the new functions/output with the rest of the code.